requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-experimental
chromadb
sentence-transformers
huggingface_hub
python-dotenv
beautifulsoup4